Checkpoint, working formatter for 5 initial test cases
So far I'm not doing anything interesting with the formatting aside from
proving that I understand how to traverse a subset of the Python AST.
This is the first time I've done something like this so its a worthwhile
exercise all on its own.

I've included the first hard test, #6, which will include whitespace
formatting of large function arguments. This is the first place where my
formatter will start getting interesting.