Add logic for formatting long docstrings.
It doesn't handle failure to correctly break lines yet, but it does
properly indent which is a bit of a feat. Turned out my test wasn't
quite valid since I wasn't using tabs in my expected output as I should
have, so I fixed that. I also fixed the diff to make it a little easier
to tell when whitespace is wrong between output and expected.