Rework indentation logic to be more consistent.
I was hitting some problems with how I had indentation logic scattered throughout. Think of this as a cleanup. Not all the tests pass yet because I destroyed my ability to do post-facto comments. I don't miss that because what I had was broken in many ways and really just got a single test to pass rather than be a solution that actually solved the problem.