Rework comments to correctly group complex comments
There were several problems here and the rework is extensive. I've changed
the comment logic so that it assumes we are always traversing the AST
in order such that we can just keep a simple index of the comments that
have been used. So far I've been able to make some simple changes to the
ordering things are calculated to make this true, and I think I can continue
to do that in the future.

The biggest problem here is that there is apparently a token that is sometimes
used to mark comments but it's not public in the 'token' module like other
token names. It's...odd. I think it might be a bug in Python. I'd like to
look in to that at some point.

For now I've added a new property, 'dedent', to comments, that seems to use
this strange token to indicate the difference between comments that happen
immediately after a dedent and ones that done. I can then use this to
identify comments that belong as tailing ends of a body vs comments that
are the start of the next line after a body.