Fix formatting of large strings.
This uses the new broken-out strings module to add a new strategy for strings
that need newlines in them because they are too long. The strategy looks
for spaces between words to allow for breaking the string at a reasonable
point.

This also treats docstrings as different from regular strings. We don't
want to format docstrings because they may reasonably include ASCII art.