Handle mutli-line strings with newlines in them.
This is a bit of a hack. I add spaces around newlines so that
the spaces stategy will be applied and then just remove the extra
spaces I added in a post-processing step.

It works correctly though.

I also fixed the algorithm for doing kwargs horizontally vs vertically
by taking into account just the last line since. Previously it would
consider the entirety of the value-under-format instead of just
considering the last line of the value-under-format. This means if
you call, say "foo".format(bar) and foo is really large you'll always
vertically align bar. That's wrong.