Fix more string formatting, this time, newline breaks.
The issue is that I didn't like the way string breaks were
working with embedded newlines. We can make them so much prettier
by aggressively preferring to break on newlines. We now greedily
group newlines together then break after them in string constants.

In doing this I introduced a new concept of "reserving" whitespace
on a given line. This only makes sense in Python since whitespace
is significant. When we are serializing the AST we can, in some
situations, know how much space has already been consumed by the
line for something like a return statement. We reserve that in
the context so that parts of the serialization that makes decisions
like string formatting can take it into account when adding line
breaks.