Add support for pretty-printing large arg lists.
I got sick of how big the base module was getting so I broke it apart
into a 'functions' module for handling logic around pretty-printing
functions. In this case, just the args.

In order to make this work I had to add a new value to the context:
the format_value function. It's not ideal, but otherwise I get a circular
dependency because nearly everything depends on format_value in order
to get its work done recursively. Passing the func through on the
context avoids this circular dependency problem in exchange for
muddying what a context is.

I also created a new 'alignment' module that will be used to
align things on a particular character. Read: any sufficiently
large comprehensions or lists of parameters.