Rework argument alignment in vertical cases.
This also adds a bunch of type hinting to places I was looking at.

This is pretty significant. I've added a README with the details
of what I've decided to do. I don't think it's complete, there's still
some cases that aren't handled properly and its a very complex feature so
it's going to take time to shake out all the bugs.

But I'm convinced this is a better direction and I've got something
working so I'm going to commit it.

As part of this I've decided the best way to handle different
strategies for laying out code is to bail as soon as we know that
a particular strategy won't work by raising an error. The parent
formatter will then have to handle the error intelligently and
try different strategies until it runs out of ideas or it is
successful.